requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
anyio==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
auto-py-to-exe==2.18.2
Babel==2.9.1
beautifulsoup4==4.11.1
bidict==0.22.0
bottle==0.12.19
bottle-websocket==0.2.9
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
connexion==2.14.0
cssselect2==0.5.0
cv==1.0.0
cx-Freeze==6.10
cx-Logging==3.0
cycler==0.11.0
decorator==5.1.1
dnspython==2.2.1
docopt==0.6.2
dp==1.0
Eel==0.12.4
email-validator==1.1.3
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==1.0.1
fonttools==4.31.2
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.14.7
httpx==0.22.0
hypercorn==0.13.2
hyperframe==6.0.1
icecream==2.1.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipaddress==1.0.23
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
Jinja2==3.0.3
jsonschema==4.4.0
kiwisolver==1.4.4
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.5.2
mongoengine==0.24.1
more-itertools==8.12.0
multidict==6.0.2
nbformat==5.3.0
networkx==2.8.5
packaging==21.3
pdfcrowd==5.5.0
pefile==2021.9.3
phonenumbers==8.12.45
pigar==0.10.0
pipreqs==0.4.11
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
psycopg2==2.9.3
py==1.11.0
py-pspdfkit==0.2.5
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydyf==0.1.2
Pygments==2.11.2
pyinstaller==5.0
pyinstaller-hooks-contrib==2022.4
pymongo==4.1.1
pyparsing==3.0.7
pyphen==0.12.0
pypiwin32==223
pypng==0.20220715.0
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.17.1
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2021.3
pyvirtualcam==0.9.1
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzbar==0.1.9
requests==2.27.1
response-report==1.1.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.30
starlette==0.19.0
swagger-ui-bundle==0.0.9
tablib==3.2.1
Telethon==1.24.0
tempora==5.0.1
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
traitlets==5.1.1
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.8
validators==0.18.2
weasyprint==54.2
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
win32-setctime==1.1.0
wkhtmltopdf==0.2
wsproto==1.1.0
WTForms==3.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.7.2
zc.lockfile==2.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
